requirements.txt
Django==1.8.4
django-compat==1.0.7
psycopg2==2.6.1
git+git://github.com/darklow/django-suit.git
django-localflavor-au-tjh==1.1.2
numpy==1.9.2
scipy==0.16.0
pandas==0.16.2
beautifulsoup4==4.4
git+git://github.com/ITCase-django/django-tinymce-4
django-mptt>=0.7.2
six>=1.5.2
django-flat-theme
django-page-cms
git+git://github.com/smacker/django-filebrowser-no-grappelli.git#egg=django-filebrowser
django-taggit
django-haystack
django-authority
whoosh
Markdown
djangorestframework
git+git://github.com/riklaunim/django-ckeditor.git#egg=django_ckeditor_updated-master
scikit-learn
uwsgi
django_cron
tornado
pexpect
django-import-export
tablib
diff_match_patch
Pillow
python-dateutil